// src/types/academic.ts
export interface Materia {
    id: number;
    nombre: string;
    codigo: string;
    descripcion: string;
    creditos: number;
    profesor: number;
    profesor_detail?: {
        username: string;
        email: string;
        first_name: string;
        last_name: string;
        role: string;
        curso?: number;
        curso_detail?: Curso;
    };
}

export interface Estudiante {
    id: number;
    username: string;
    email: string;
    first_name: string;
    last_name: string;
    role: string;
    curso?: number;
    curso_detail?: Curso;
}

export interface Curso {
    id: number;
    nombre: string;
    nivel: string; // 'PRIMARIA' | 'SECUNDARIA'
    materias: number[];
}